revert changes made to express-sessions stuff
